[FEAT] improve and refactor my orders pages
