+++ src/components/MyOrders/index.js
@@ -1,224 +1,17 @@
-import React, { useState } from 'react'
-import Skeleton from 'react-loading-skeleton'
-import {
-  MyOrders as MyOrdersController,
-  useEvent,
-  useLanguage,
-  useOrder,
-  useUtils
-} from 'ordering-components'
-import {
-  MyOrdersContainer,
-  MyOrdersTitle,
-  ActiveOrders,
-  Card,
-  Map,
-  Content,
-  Logo,
-  PastLogo,
-  BusinessInformation,
-  OpenOrder,
-  Price,
-  OrdersPast,
-  IndividualOrderPast,
-  OrderPastContent,
-  WrapperBusinessTitle,
-  Reorder,
-  ImageContainer,
-  SkeletonOrder,
-  SkeletonCard,
-  SkeletonMap,
-  SkeletonInformation,
-  SkeletonPastContent,
-  SkeletonContent,
-  SkeletonText,
-  SkeletonOrdersPast,
-  SkeletonReorder
-} from './styles'
+import React from 'react'
+import { Container } from './styles'
 
 import { ProfileOptions } from '../UserProfileForm/ProfileOptions'
-import { AutoScroll } from '../AutoScroll'
-import { Button } from '../../styles/Buttons'
-import { Tabs } from '../../styles/Tabs'
-import emptyPastOrders from '../../../template/assets/empty-past-orders.svg'
-import emptyActiveOrders from '../../../template/assets/empty-active-orders.svg'
+import { OrdersOption } from '../OrdersOption'
 
-export const MyOrdersUI = (props) => {
-  const { activeOrders, previousOrders } = props
-  const [, t] = useLanguage()
-  const [, { reorder }] = useOrder()
-  const [events] = useEvent()
-  const [{ parsePrice }] = useUtils()
-  const googleMapKey = 'AIzaSyDX5giPfK-mtbLR72qxzevCYSUrbi832Sk'
-  const getGoogleMapImage = ({ lat, lng }) => {
-    return `https://maps.googleapis.com/maps/api/staticmap?size=500x190&center=${lat},${lng}&zoom=17&scale=2&maptype=roadmap&&markers=icon:https://res.cloudinary.com/ditpjbrmz/image/upload/f_auto,q_auto,w_45,q_auto:best,q_auto:best/v1564675872/marker-customer_kvxric.png%7Ccolor:white%7C${lat},${lng}&key=${googleMapKey}`
-  }
-
-  const activeOrdersSorted = activeOrders?.orders?.sort((a, b) => new Date(b.created_at) - new Date(a.created_at)) ?? []
-
-  const [reorderLoading, setReorderLoading] = useState(false)
-  const [orderID, setOrderID] = useState(null)
-
-  const handleReorder = async (orderId) => {
-    setReorderLoading(true)
-    setOrderID(orderId)
-    try {
-      const { error, result } = await reorder(orderId)
-      if (!error) {
-        events.emit('go_to_page', { page: 'checkout', params: { cartUuid: result.uuid } })
-      }
-      setReorderLoading(false)
-    } catch (err) {
-      console.log(err)
-      setReorderLoading(false)
-    }
-  }
-
-  const handleGoToPage = (data) => {
-    events.emit('go_to_page', data)
-  }
-
+export const MyOrders = () => {
   return (
     <>
       <ProfileOptions value='My Orders' />
-      <MyOrdersContainer>
-        {
-          <>
-            <MyOrdersTitle>
-              <h1>{t('ACTIVE_ORDERS', 'Active Orders')}</h1>
-            </MyOrdersTitle>
-            {!activeOrders.loading && activeOrders.orders.length === 0 ? (
-              <ImageContainer>
-                <img src={emptyActiveOrders} alt='empty-active-orders' />
-              </ImageContainer>
-            ) : (
-              <ActiveOrders id='container'>
-                <Tabs id='orders'>
-                  <AutoScroll container='container' categories='orders' loading={activeOrders.loading}>
-                    {!activeOrders.loading ? activeOrdersSorted.map(order => (
-                      <Card key={order.id}>
-                        <Map>
-                          <img src={getGoogleMapImage(order?.business?.location)} alt='google-maps-img' />
-                        </Map>
-                        <Content>
-                          <Logo>
-                            <img src={order.business?.logo} alt='business-logo' />
-                          </Logo>
-                          <BusinessInformation>
-                            <h2>{order.business.name}</h2>
-                            <p>Order No. {order.id}</p>
-                            <p>{order.created_at}</p>
-                          </BusinessInformation>
-                          <Price>
-                            <h2>
-                              {parsePrice(order.products.reduce((acc, cur) => acc + cur.price, 0))}
-                            </h2>
-                            {order.status === 0 && (
-                              <p>{t('ORDER_PENDING', 'pending')}</p>
-                            )}
-                          </Price>
-                        </Content>
-                        <OpenOrder>
-                          <Button color='primary' onClick={() => handleGoToPage({ page: 'order_detail', params: { orderId: order.id } })}>
-                            {t('OPEN_ORDER', 'Open order')}
-                          </Button>
-                        </OpenOrder>
-                      </Card>
-                    )) : (
-                      <SkeletonOrder>
-                        {[...Array(3)].map((item, i) => (
-                          <SkeletonCard key={i}>
-                            <SkeletonMap>
-                              <Skeleton />
-                            </SkeletonMap>
-                            <SkeletonContent>
-                              <div>
-                                <Skeleton width={70} height={70} />
-                              </div>
-                              <SkeletonText>
-                                <Skeleton width={100} />
-                                <Skeleton width={80} />
-                                <Skeleton width={120} />
-                              </SkeletonText>
-                            </SkeletonContent>
-                          </SkeletonCard>
-                        ))}
-                      </SkeletonOrder>
-                    )}
-                  </AutoScroll>
-                </Tabs>
-              </ActiveOrders>
-            )}
-          </>
-        }
-        {
-          <>
-            <MyOrdersTitle>
-              <h1>{t('ORDERS_PAST', 'My Orders Past')}</h1>
-            </MyOrdersTitle>
-            {!previousOrders.loading && previousOrders.orders.length === 0 ? (
-              (
-                <ImageContainer>
-                  <img src={emptyPastOrders} alt='empty-past-orders' />
-                </ImageContainer>
-              )
-            ) : (
-              <OrdersPast>
-                {!previousOrders.loading ? previousOrders.orders.map(order => (
-                  <IndividualOrderPast key={order.id}>
-                    <OrderPastContent>
-                      <PastLogo>
-                        <img src={order.business?.logo} alt='business-logo' />
-                      </PastLogo>
-                      <BusinessInformation>
-                        <WrapperBusinessTitle>
-                          <h2>{order.business.name}</h2>
-                        </WrapperBusinessTitle>
-                        <p>{order.created_at}</p>
-                        <p name='view_order' onClick={() => handleGoToPage({ page: 'order_detail', params: { orderId: order.id } })}>{t('MOBILE_FRONT_BUTTON_VIEW_ORDER', 'View order')}</p>
-                      </BusinessInformation>
-                    </OrderPastContent>
-                    <Reorder>
-                      {(order.status === 1 || order.status === 11) && (
-                        <p>{t('ORDER_COMPLETED', 'Complete')}</p>
-                      )}
-                      <Button color='primary' onClick={() => handleReorder(order.id)} disabled={reorderLoading}>{orderID === order.id && reorderLoading ? t('LOADING', 'Loading...') : t('REORDER', 'Reorder')}</Button>
-                    </Reorder>
-                  </IndividualOrderPast>
-                )) : ([...Array(3)].map((item, i) => (
-                  <SkeletonOrdersPast key={i}>
-                    <SkeletonPastContent>
-                      <SkeletonInformation>
-                        <div>
-                          <Skeleton width={70} height={70} />
-                        </div>
-                        <SkeletonText>
-                          <Skeleton width={100} />
-                          <Skeleton width={120} />
-                          <Skeleton width={80} />
-                        </SkeletonText>
-                      </SkeletonInformation>
-                      <SkeletonReorder>
-                        <Skeleton />
-                        <Skeleton />
-                      </SkeletonReorder>
-                    </SkeletonPastContent>
-                  </SkeletonOrdersPast>
-                )))}
-              </OrdersPast>
-            )}
-          </>
-        }
-      </MyOrdersContainer>
+      <Container>
+        <OrdersOption activeOrders />
+        <OrdersOption />
+      </Container>
     </>
   )
-}
-
-export const MyOrders = (props) => {
-  const MyOrdersProps = {
-    ...props,
-    UIComponent: MyOrdersUI
-  }
-
-  return <MyOrdersController {...MyOrdersProps} />
 }
